Add image link formatting
Image links are complex links using a picture instead of a text caption.
They have the following format:
  [[URL|{{file|caption}}]]
where no }} or ]] are allowed inside curly braces. The previous code had
a bug when alone } or ] broke highlighting of such links. This patch
solves this bug.
Additionally I introduced a separate highlighting for the symbols ],},
and | when they are used for links formatting.
Multiline links are allowed.