Define a separate dokuwikiTodo for FIXME and DELETEME
These have been lumped together with the smileys, but serve different purposes: The latter convey emotions, but the former alert maintainers to problems (and as such, users may want to emphasize those, e.g. with boldface and a glaring yellow background).
As FIXME and DELETEME are just alphabetic characters, :syn keyword is used (this wouldn't be possible for the smileys) - this also prevents inadvertent matching inside other text (PREFIXMEDIATION).
Keep the default linking to the Todo group for smileys to maintain backwards compatibility - users have to customize if they want to distinguish both.